+++ bbtest/helpers/restful_api.rb
@@ -5,11 +5,16 @@ class RestfulAPI
 
   def call(http_method, url, data = {})
     case http_method.downcase
-    when "get";    return get(url)
-    when "post";   return post(url, data)
-    when "patch";  return patch(url, data)
-    when "delete"; return post(url)
-    else;          raise "undefined method #{http_method.upcase}"
+    when "get"
+      return get(url)
+    when "post"
+      return post(url, data)
+    when "patch"
+      return patch(url, data)
+    when "delete"
+      return post(url)
+    else
+      raise "undefined method #{http_method.upcase}"
     end
   end