+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 - First release
 - ZMQ distributed queue
+- Debian packaging